#38 Extract logic of CreateTablesMojo
