+++ src/main/java/com/jcabi/dynamodb/core/package-info.java
@@ -34,14 +34,16 @@
  *
  * <pre>
  * {@code
- * &lt;depedency&gt;
- *   &lt;groupId&gt;com.jcabi&lt;/groupId&gt;
- *   &lt;artifactId&gt;jcabi-dynamodb-maven-plugin&lt;/artifactId&gt;
- * &lt;/dependency&gt;
+ * <depedency>
+ *   <groupId>com.jcabi</groupId>
+ *   <artifactId>jcabi-dynamodb-maven-plugin</artifactId>
+ * </dependency>
  * }
  * </pre>
  * @author Igor Piddubnyi (igor.piddubnyi@gmail.com)
  * @version $Id$
  * @since 0.8
+ * @todo #38:30min Logic of the plugin should be extracted from maven mojos to
+ *  this package to allow directly running local dynamodb.
  */
 package com.jcabi.dynamodb.core;
